feat(nextjs): Simplify auth().protect() types
