+++ packages/nextjs/src/server/getAuth.ts
@@ -1,5 +1,5 @@
 import type { Organization, Session, User } from '@clerk/backend';
-import type { SignedInAuthObject, SignedOutAuthObject } from '@clerk/backend/internal';
+import type { AuthObject } from '@clerk/backend/internal';
 import {
   AuthStatus,
   constants,
@@ -13,7 +13,7 @@ import { decodeJwt } from '@clerk/backend/jwt';
 import { withLogger } from '../utils/debugLogger';
 import { API_URL, API_VERSION, SECRET_KEY } from './constants';
 import { getAuthAuthHeaderMissing } from './errors';
-import type { AuthObjectWithoutResources, RequestLike } from './types';
+import type { RequestLike } from './types';
 import { getAuthKeyFromRequest, getCookie, getHeader, injectSSRStateIntoObject } from './utils';
 
 export const createGetAuth = ({
@@ -24,10 +24,7 @@ export const createGetAuth = ({
   debugLoggerName: string;
 }) =>
   withLogger(debugLoggerName, logger => {
-    return (
-      req: RequestLike,
-      opts?: { secretKey?: string },
-    ): AuthObjectWithoutResources<SignedInAuthObject> | AuthObjectWithoutResources<SignedOutAuthObject> => {
+    return (req: RequestLike, opts?: { secretKey?: string }): AuthObject => {
       const debug = getHeader(req, constants.Headers.EnableDebug) === 'true';
       if (debug) {
         logger.enable();
