+++ packages/nextjs/src/server/types.ts
@@ -1,4 +1,4 @@
-import type { AuthObject, OptionalVerifyTokenOptions } from '@clerk/backend/internal';
+import type { OptionalVerifyTokenOptions } from '@clerk/backend/internal';
 import type { MultiDomainAndOrProxy } from '@clerk/types';
 import type { IncomingMessage } from 'http';
 import type { NextApiRequest } from 'next';
@@ -20,5 +20,3 @@ export type WithAuthOptions = OptionalVerifyTokenOptions &
   };
 
 export type NextMiddlewareResult = Awaited<ReturnType<NextMiddleware>>;
-
-export type AuthObjectWithoutResources<T extends AuthObject> = Omit<T, 'user' | 'organization' | 'session'>;
